Add some casts.
svn path=/trunk/; revision=51122

+++ ui/gtk/addr_resolution_dlg.c
@@ -170,16 +170,16 @@ eth_hash_to_texbuff(gpointer key, gpointer value, gpointer user_data)
 	gint64 eth_as_gint64 = *(gint64*)key;
 	hashether_t* tp = (hashether_t*)value;
 
-	g_snprintf(string_buff, ADDRESS_STR_MAX, "%.2X:%.2X:%.2X:%.2X:%.2X:%.2X Status: %u %s %s\n", 
-		eth_as_gint64>>40&0xff,
-		eth_as_gint64>>32&0xff,
-		(eth_as_gint64>>24)&0xff,
-		(eth_as_gint64>>16)&0xff, 
-		(eth_as_gint64>>8)&0xff, 
-		eth_as_gint64&0xff,
-		tp->status,
-		tp->hexaddr,
-		tp->resolved_name);
+	g_snprintf(string_buff, ADDRESS_STR_MAX, "%.2X:%.2X:%.2X:%.2X:%.2X:%.2X Status: %u %s %s\n",
+		   (guint8)(eth_as_gint64>>40&0xff),
+		   (guint8)(eth_as_gint64>>32&0xff),
+		   (guint8)((eth_as_gint64>>24)&0xff),
+		   (guint8)((eth_as_gint64>>16)&0xff),
+		   (guint8)((eth_as_gint64>>8)&0xff),
+		   (guint8)(eth_as_gint64&0xff),
+		   tp->status,
+		   tp->hexaddr,
+		   tp->resolved_name);
 	gtk_text_buffer_insert_at_cursor (buffer, string_buff, -1);
 
 }
@@ -204,14 +204,14 @@ wka_hash_to_texbuff(gpointer key, gpointer value, gpointer user_data)
 	gchar *name = (gchar *)value;
 	gint64 eth_as_gint64 = *(gint64*)key;
 
-	g_snprintf(string_buff, ADDRESS_STR_MAX, "%.2X:%.2X:%.2X:%.2X:%.2X:%.2X  %s\n", 
-		eth_as_gint64>>40&0xff,
-		eth_as_gint64>>32&0xff,
-		(eth_as_gint64>>24)&0xff,
-		(eth_as_gint64>>16)&0xff, 
-		(eth_as_gint64>>8)&0xff, 
-		eth_as_gint64&0xff,
-		name);
+	g_snprintf(string_buff, ADDRESS_STR_MAX, "%.2X:%.2X:%.2X:%.2X:%.2X:%.2X  %s\n",
+		   (guint8)(eth_as_gint64>>40&0xff),
+		   (guint8)(eth_as_gint64>>32&0xff),
+		   (guint8)((eth_as_gint64>>24)&0xff),
+		   (guint8)((eth_as_gint64>>16)&0xff),
+		   (guint8)((eth_as_gint64>>8)&0xff),
+		   (guint8)(eth_as_gint64&0xff),
+		   name);
 	gtk_text_buffer_insert_at_cursor (buffer, string_buff, -1);
 
 }
